+++ NEWS.md
@@ -1,3 +1,7 @@
+# shiny.esreglog 0.5.7
+
+Added admin tags for users database and relevant function calls to set and remove admin status
+
 # shiny.esreglog 0.5.6
 
 Allowed server to specify a username rather than always using the email as the username for users who sign in with Microsoft.